Moved description picker to new location.
