+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/details/description-picker.js
